requirements.txt
langchain>=0.3
langchain-openai>=0.2
langchain-community>=0.3
langgraph>=0.2
openai>=1.0
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
qdrant-client>=1.9
pydantic>=2.0
python-multipart>=0.0.9
langchain-text-splitters>=0.3.0